Fixed a potential deadlock in the has_work() function of the input_node_data_writer which could happen if the data stream ended precisely at the end of the time_interval.
git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@990 5736e3ea-d426-0410-8aaf-93e336ac3372

+++ include/input_node_data_writer.h
@@ -90,7 +90,7 @@ private:
   int bits_per_sample;
   int fftsize;
   int time_fft;
-  int64_t time_per_byte;
+  double time_per_byte;
   int integration_time;
 
   /// The queue storing all the delays
